feat: ensure DeleteCommandByIdController will return 400 if LoadCommandById returns null

+++ src/presentation/controllers/command/delete-command-by-id/delete-command-by-id-controller.ts
@@ -16,7 +16,10 @@ export class DeleteCommandByIdController implements Controller {
 
   async handle(httpRequest: HttpRequest): Promise<HttpResponse> {
     try {
-      await this.loadCommandById.loadById(httpRequest.params.commandId);
+      const command = await this.loadCommandById.loadById(httpRequest.params.commandId);
+      if (!command) {
+        return badRequest(new InvalidParamError('commandId'));
+      }
       const deleted = await this.deleteCommandById.deleteById(httpRequest.params.commandId);
       return deleted ? noContent() : badRequest(new InvalidParamError('commandId'));
     } catch (err) {
